Drop `aria-controls` from control elements that target multiple elements
According to [Bootstrap
documentation](https://getbootstrap.com/docs/4.0/components/collapse/#accessibility)
it should be added when the control element collapses a single target:

> If your control element is targeting a single collapsible element –
i.e. the data-target attribute is pointing to an id selector – you
should add the aria-controls attribute to the control element,
containing the id of the collapsible element.